docker/requirements.txt
fastapi
uvicorn[standard]
assemblyai
twilio
python-dotenv
ngrok
aiohttp
requests
vosk
websockets
google-cloud-texttospeech
speechrecognition
numpy
librosa
pydub
chromadb
psycopg2-binary
langgraph
langchain-core
langchain-openai
fuzzywuzzy
protobuf==5.29.3
redis
python-multipart
boto3

--extra-index-url https://download.pytorch.org/whl/cpu
torch
torchaudio
sentence_transformers
